requirements.txt
amqp==2.6.1
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
blinker==1.4
Brotli==1.0.9
cached-property==1.5.2
celery==4.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.2
Deprecated==1.2.13
Django==3.2
django-cacheops==6.0
django-cors-headers==3.11.0
django-countries==7.3.2
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
django-multiselectfield==0.1.12
django-oidc-provider==0.7.0
django-rest-passwordreset==1.2.1
django-username-email==2.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
enum-compat==0.0.3
funcy==1.17
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
kaitaistruct==0.9
kombu==4.6.11
outcome==1.1.0
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.14.1
pydivert==2.1.0
pyjwkest==1.4.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.0
requests==2.27.1
selenium==4.1.3
selenium-wire==4.6.5
sentry-sdk==1.5.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
vine==1.3.0
wcwidth==0.2.5
webdriver-manager==3.5.4
wrapt==1.14.0
wsproto==1.1.0
zope.event==4.5.0
zope.interface==5.4.0
zstandard==0.17.0
imap-tools==0.56.0
undetected-chromedriver
pandas
pyvirtualdisplay
xvfbwrapper
tk
selenium-stealth
kameleo.local-api-client